Move all packed field layout configuration code to TupleLayout class
